lots of new tests and better search

+++ app/modules/common/schemas/gear-schema.ts
@@ -9,7 +9,7 @@ const gearSchema = z.object({
   link: z.string().nullable().default(null),
   notes: z.string().nullable().default(null),
   consumable: z.boolean().default(false),
-  price: z.number().int().nullable().default(null),
+  price: z.number().nullable().default(null),
   currency: z.nativeEnum(Currency),
 });
